+++ HaremKing.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HaremKingv3", "HaremKingv3\HaremKingv3.csproj", "{30B388AA-13A3-40DE-9B9F-0EF500F0F53C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HaremKing", "HaremKing\HaremKing.csproj", "{30B388AA-13A3-40DE-9B9F-0EF500F0F53C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
